Add Ray and feature to switch between f32/f64

+++ src/geometry/point3.rs
@@ -1,6 +1,6 @@
 use geometry::vector3 as vec3;
 use na::{Point2, Point3, Vector3};
-use types::{pmax, pmin, ElemType, HasNan, Point3f};
+use types::{pmax, pmin, ElemType, Float, HasNan, Point3f};
 
 pub fn to_pt2<T: ElemType>(p: &Point3<T>) -> Point2<T> {
     assert!(!p.has_nan());
@@ -11,7 +11,7 @@ pub fn to_vec3<T: ElemType>(p: &Point3<T>) -> Vector3<T> {
     Vector3::new(p[0], p[1], p[2])
 }
 
-pub fn lerp(t: f64, p0: &Point3f, p1: &Point3f) -> Point3f {
+pub fn lerp(t: Float, p0: &Point3f, p1: &Point3f) -> Point3f {
     vec3::to_pt3(&((1.0 - t) * p0.coords + t * p1.coords))
 }
